Crate gatelogue_types

Rust utility library for using Gatelogue data in Rust projects. It will load the database for you to access via ORM or raw SQL.

Installation

Add to your Cargo.toml:

gatelogue-types = { version = "3", features = [...] }

# To import directly from the repository:
gatelogue-types = { git = "https://github.com/mrt-map/gatelogue", package = "gatelogue-types", features = [...] }

Features as denoted by ... include reqwest_get, surf_get, ureq_get (to download with reqwest, surf and ureq respectively) and bundled (to bundle SQLite).

Usage

To retrieve the data:

use gatelogue_types::GD;
let gd = GD::reqwest_get_with_sources().await?; // with sources, requires `reqwest_get` feature
let gd = GD::reqwest_get_no_sources().await?; // no sources, requires `reqwest_get` feature
let gd = GD::surf_get_with_sources().await?; // with sources, requires `surf_get` feature
let gd = GD::surf_get_no_sources().await?; // no sources, requires `surf_get` feature
let gd = GD::ureq_get_with_sources()?; // with sources, requires `ureq_get` feature
let gd = GD::ureq_get_no_sources()?; // no sources, requires `ureq_get` feature

Using the ORM does not require SQL and makes for generally clean code. However, doing this is very inefficient as each attribute access is one SQL query.

use gatelogue_types::AirAirport;
for airport in gd.nodes_of_type::<AirAirport>()? {
    for gate in airport.gates(&gd)? {
        println!("Airport {:?} has gate {:?}", airport.code(&gd)?, gate.code(&gd)?)
    }
}

Querying the underlying SQLite database directly with rusqlite is generally more efficient and faster. It is also the only way to access the *Source tables, if you retrieved the database with those.

gd.0.prepare("SELECT A.code, G.code FROM AirGate G INNER JOIN AirAirport A ON G.airport = A.i")?
    .query_map((), |row| {
        println!("Airport {:?} has gate {:?}", row.get::<_, String>(0)?, row.get::<_, String>(1)?);
        Ok(())
    })?;

Macros

from_sql_for_enum

get_aircraft_column

get_column

get_derived_vec

get_set

node_type

Structs

AirAirline

AirAirport

AirFlight

AirGate

Aircraft

BusBerth

BusCompany

BusConnection

BusLine

BusStop

GD

Proximity

RailCompany

RailConnection

RailLine

RailPlatform

RailStation

SeaCompany

SeaConnection

SeaDock

SeaLine

SeaStop

SpawnWarp

Town

Enums

AirMode

AnyLocatedNode

AnyNode

BusMode

Error

RailMode

Rank

SeaMode

WarpType

World

Constants

URL

URL_NO_SOURCES

Traits

LocatedNode

Node

Type Aliases

ID

Result